refactor: apply defaults by pure way
