+++ src/core/core.layouts.defaults.js
@@ -1,0 +1,11 @@
+export function applyLayoutsDefaults(defaults) {
+  defaults.set('layout', {
+    autoPadding: true,
+    padding: {
+      top: 0,
+      right: 0,
+      bottom: 0,
+      left: 0
+    }
+  });
+}
